Crate market

Market for EACs

Modules

sale

sale_views

Structs

Contract

Struct for storing various information about market state

Enums

StorageKey

Helper structure to for keys of the persistent collections.

Constants

CCC

NO_DEPOSIT

PROCESS_ASK

STORAGE_PER_SALE

Functions

new

storage_balance_of

Return how much storage an account has paid for

storage_deposit

Allows users to deposit storage. This is to cover the cost of storing sale objects on the contract

storage_minimum_balance

Return the minimum storage for 1 sale

storage_withdraw

Allows users to withdraw any excess storage that they’re not using

Type Definitions

ContractAndId

SalePriceInYoctoNear