Struct market::Contract

pub struct Contract {
    pub owner_id: AccountId,
    pub asks: UnorderedMap<ContractAndId, Ask>,
    pub bids: UnorderedMap<ContractAndId, Bid>,
    pub asks_id: u128,
    pub bids_id: u128,
    pub asks_by_owner_id: LookupMap<AccountId, UnorderedSet<ContractAndId>>,
    pub bids_by_owner_id: LookupMap<AccountId, UnorderedSet<ContractAndId>>,
    pub storage_deposits: LookupMap<AccountId, Balance>,
}

Struct for storing various information about market state

Fields
owner_id: AccountIdasks: UnorderedMap<ContractAndId, Ask>

To keep track of the ask, we map the ContractAndTokenId to a Sale. the ContractAndTokenId is the unique identifier for every sale. It is made up of the contract ID + DELIMITER + UUIDv4

bids: UnorderedMap<ContractAndId, Bid>

Same with bids

asks_id: u128

Current id for asks

bids_id: u128

Current id for bids

asks_by_owner_id: LookupMap<AccountId, UnorderedSet<ContractAndId>>

Keep track of all the Ask IDs for every account ID

bids_by_owner_id: LookupMap<AccountId, UnorderedSet<ContractAndId>>

Keep track of all the Bids IDs for every account ID

storage_deposits: LookupMap<AccountId, Balance>

Keep track of the storage that accounts have paid

Implementations

impl Contract

pub fn internal_place_ask(&mut self, ask: Ask) -> Ask

pub fn internal_place_bid(&mut self, bid: Bid) -> Bid

pub fn internal_remove_ask(&mut self, id: ContractAndId) -> Ask

Internal method for removing a ask from the market. This returns the previously removed object

pub fn internal_remove_bid(&mut self, id: ContractAndId) -> Bid

Internal method for removing a bid from the market. This returns the previously removed object

impl Contract

pub fn place_ask(&mut self, amount: Balance, conditions: U128) -> Promise

Place Ask position on a market

Arguments

amount: FT amount

conditions: For how much $NEAR

returns: Promise to transfer

pub fn place_bid(&mut self, amount: Balance, conditions: U128) -> Bid

Place Bid on a market

Arguments

amount: How much to buy

conditions: For how much $NEAR

returns: Bid that was created

pub fn remove_position(&mut self, id: ContractAndId, position: Position)

Removes ask from the market.

pub fn cancel_position(
&mut self,
id: ContractAndId,
position: Position
) -> Promise

Cancels position from “user-mode”

Arguments

id: ContractAndId of position

position: Position type

returns: Promise that transfers FT or $NEAR

pub fn update_price(
&mut self,
id: ContractAndId,
price: U128,
position: Position
)

Updates the price for a sale on the market

pub fn process_bid(
&mut self,
ask_id: ContractAndId,
bid_id: ContractAndId
) -> Promise

Process Bid. Can be used only by owner (for back-end)

Arguments

ask_id: Ask

bid_id: Bid

returns: Promise

pub fn direct_ask_sell(&mut self, id: ContractAndId) -> Promise

Make direct sell of ask (by user)

Arguments

id: ContractAndId

returns: Promise

pub fn resolve_position(&mut self, id: ContractAndId, position: Position)

impl Contract

Views. Used to retrieve information from market

pub fn get_supply_asks(&self) -> U64

Returns the number of sales the marketplace has up (as a string)

pub fn get_supply_bids(&self) -> U64

Returns the number of sales the marketplace has up (as a string)

pub fn get_supply_by_owner_id(
&self,
account_id: AccountId,
position: Position
) -> U64

Get number of Positions by [AccountId]

Arguments

account_id:

position: Position

returns: U64

pub fn get_asks_by_owner_id(
&self,
account_id: AccountId,
from: Option<u128>,
limit: Option<u64>
) -> Vec<Ask>

Get paginated list of Asks for specified user

Arguments

account_id: [AccountId]

from:

limit:

returns: Vec<Ask, Global>

pub fn get_bids(&self, from: Option<U128>, limit: Option<u64>) -> Vec<Bid>

Get paginated list of Bids

Arguments

from:

limit:

returns: List of bids

pub fn get_asks(&self, from: Option<U128>, limit: Option<u64>) -> Vec<Ask>

Get paginated list of Asks

Arguments

from:

limit:

returns: List of asks

pub fn get_bids_by_owner_id(
&self,
account_id: AccountId,
from: Option<U128>,
limit: Option<u64>
) -> Vec<Bid>

Get paginated list of Bids for specified user

Arguments

account_id: [AccountId]

from:

limit:

returns: Vec<Ask, Global>

pub fn get_ask(&self, id: &ContractAndId) -> Option<Ask>

Get a ask information for a given unique ask ID (contract + DELIMITER + uuid)

pub fn get_bid(&self, id: &ContractAndId) -> Option<Bid>

Get a bid information for a given unique bid ID (contract + DELIMITER + uuid)

impl Contract

pub fn new(owner_id: AccountId) -> Self

pub fn storage_deposit(&mut self, account_id: Option<AccountId>)

Allows users to deposit storage. This is to cover the cost of storing sale objects on the contract

pub fn storage_withdraw(&mut self)

Allows users to withdraw any excess storage that they’re not using

pub fn storage_minimum_balance(&self) -> U128

Return the minimum storage for 1 sale

pub fn storage_balance_of(&self, account_id: AccountId) -> U128

Return how much storage an account has paid for